Adding nuess0r keymap for the ctrl_m
ctrl-M is a replacement controller for IBM Model M keyboards.

The default keymap in QMK is minimal and functions as a starting point to define for your own.

The nuess0r keymap adds a lot of useful additions compared to a stock Model M like windows and
multimedia keys, mouse wheel support etc.

This is the pre-flashed keymap when you buy a ctrl-M kit.

Details are described here:
https://brain4free.org/wiki/doku.php/elektronik:ctrl_m#nuess0r_keymap